finish

+++ src/main/kotlin/com/milsondev/downloaduploadfiles/service/MyFileService.java
@@ -1,10 +1,14 @@
 package com.milsondev.downloaduploadfiles.service;
 
 import com.milsondev.downloaduploadfiles.api.Category;
+import com.milsondev.downloaduploadfiles.db.entity.FileContent;
 import com.milsondev.downloaduploadfiles.db.entity.MyFile;
+import com.milsondev.downloaduploadfiles.db.repository.FileContentRepository;
 import com.milsondev.downloaduploadfiles.db.repository.MyFileRepository;
+import com.milsondev.downloaduploadfiles.exceptions.DuplicateFileException;
 import com.milsondev.downloaduploadfiles.exceptions.FileSizeException;
 import com.milsondev.downloaduploadfiles.exceptions.MaximumNumberOfFilesExceptions;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -13,14 +17,15 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.text.DecimalFormat;
 import java.time.Instant;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.zip.CRC32;
+import java.util.zip.Checksum;
 
 @Service
 public class MyFileService {
     private final MyFileRepository myFileRepository;
+
+    private final FileContentRepository fileContentRepository;
 
     @Value("${max.file.size:1048576}")
     private long MAX_FILE_SIZE;
@@ -28,22 +33,48 @@ public class MyFileService {
     @Value("${max.number.of.files:10}")
     private long MAX_NUMBER_OF_FILES;
 
+    @Value("${delete.and.add.file:false}")
+    private boolean DELETE_AND_ADD_FILE;
+
     @Autowired
-    public MyFileService(final MyFileRepository myFileRepository) {
+    public MyFileService(final MyFileRepository myFileRepository,
+                         final FileContentRepository fileContentRepository) {
         this.myFileRepository = myFileRepository;
+        this.fileContentRepository = fileContentRepository;
     }
 
     public void saveFile(final MultipartFile file, Category category) throws IOException {
-        if(getMyFileList().size() < MAX_NUMBER_OF_FILES ){
-            if(file.getSize() <= MAX_FILE_SIZE ){
-                MyFile myFile = convertMultipartFileToMyFile(file, category);
-                myFileRepository.save(myFile);
+        if(DELETE_AND_ADD_FILE){
+            if(getMyFileList().size() < MAX_NUMBER_OF_FILES ){
+                if(file.getSize() <= MAX_FILE_SIZE ){
+                    if(!isDuplicateFile(file)){
+                        MyFile myFile = convertMultipartFileToMyFile(file, category);
+                        UUID myFileId = myFileRepository.save(myFile).getId();
+                        fileContentRepository.save(new FileContent(file.getBytes(), myFileId));
+                    } else {
+                        throw new DuplicateFileException("Error: File "+ file.getOriginalFilename() +" already uploaded");
+                    }
+                } else {
+                    throw new FileSizeException("Error: The file you're trying to upload is too too large " + formatSize(file.getSize()));
+                }
             } else {
-                throw new FileSizeException("Error: The file you're trying to upload is too too large " + formatSize(file.getSize()));
+                throw new MaximumNumberOfFilesExceptions("Error: You've reached the maximum number of allowed uploads. Please delete one to upload a new file.");
             }
-        } else {
-            throw new MaximumNumberOfFilesExceptions("Error: You've reached the maximum number of allowed uploads. Please delete one to upload a new file.");
         }
+    }
+
+    private boolean isDuplicateFile(MultipartFile file) throws IOException {
+        Long checksum = calculateChecksum(file.getBytes());
+        String originalFilename = file.getOriginalFilename();
+        String size = formatSize(file.getSize());
+        List<MyFile> arquivos = myFileRepository.findByChecksumAndOriginalFilenameAndSize(checksum, originalFilename, size);
+        return !arquivos.isEmpty();
+    }
+
+    private long calculateChecksum(final byte[] bytes) {
+        Checksum cRC32Checksum = new CRC32();
+        cRC32Checksum.update(bytes, 0, bytes.length);
+        return cRC32Checksum.getValue();
     }
 
     public MyFile convertMultipartFileToMyFile(final MultipartFile file, Category category) throws IOException {
@@ -52,18 +83,14 @@ public class MyFileService {
         myFile.setOriginalFilename(file.getOriginalFilename());
         myFile.setSize(formatSize(file.getSize()));
         myFile.setUploadDate(Instant.now());
-        myFile.setContent(file.getBytes());
         myFile.setContentType(file.getContentType());
         myFile.setCategory(category);
+        myFile.setChecksum(calculateChecksum(file.getBytes()));
         return myFile;
     }
 
     public List<MyFile> getMyFileList() {
         return myFileRepository.findAll();
-    }
-
-    public void deleteAll(){
-        myFileRepository.deleteAll();
     }
 
     public List<String> getCategories(){
@@ -88,12 +115,23 @@ public class MyFileService {
     }
 
     public void deleteFile(UUID id) {
-        myFileRepository.deleteById(id);
+        if(DELETE_AND_ADD_FILE){
+            myFileRepository.deleteById(id);
+        }
     }
 
+    @Transactional
     public MyFile downloadFile(UUID id) {
-        Optional<MyFile> OptionalMyFile = myFileRepository.findById(id);
-        return OptionalMyFile.orElse(null);
-    }
+        if(DELETE_AND_ADD_FILE){
+            Optional<MyFile> optionalMyFile = myFileRepository.findById(id);
+            if(optionalMyFile.isPresent()){
+                MyFile myFile = optionalMyFile.get();
+                FileContent fileContent = fileContentRepository.findByFileId(id);
+                myFile.setContent(fileContent.getContent());
+                return myFile;
+            }
+        }
 
+        return null;
+    }
 }
